1. include symlinks in the zip lambda package  2. remove yarn as nohoist is no longer needed.
include symlinks as that avoid duplicate packages in final lambda deployment.
e.g. if Service A dependes on library B and C, but lib B also depends on C, then you would get two copies of C in old approach.
But with symlinks, there is only one copy of C, which Service A and lib B links to.